+++ localstorage/storage.go
@@ -8,7 +8,7 @@ import (
 
 type ReadableStorage interface {
 	ListFiles(ctx context.Context, p string) ([]string, error)
-	ReadFile(ctx context.Context, p string) (io.ReadCloser, error)
+	ReadFile(ctx context.Context, p string) (io.ReadSeekCloser, error)
 }
 
 type WritableStorage interface {