+++ changelog.md
@@ -7,7 +7,7 @@
 - Implement generic environment variable fallback mechanism for flags.
 - add `CAS_VERBOSE` flag for `fetch`
 - add `--debug` to `fetch`, writes a metadata file to the backend with intermediate hashes, useful for debugging why a hash has changed
-
+- `fetch` will only download an artifact if the local version doesn't exist, or doesn't match the stored version
 
 ## [0.0.7] - 2022-12-18